Very important debugging of the Mcmc algorithm ...

+++ src/core/default.py
@@ -4,7 +4,8 @@ import torch
 
 logger_format = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
 
-tensor_scalar_type = torch.DoubleTensor
+# tensor_scalar_type = torch.DoubleTensor
+tensor_scalar_type = torch.FloatTensor
 tensor_integer_type = torch.LongTensor
 # deformation_kernel = kernel_factory.factory(kernel_factory.Type.TORCH, kernel_width=1.)
 deformation_kernel = None
